Dialogue Manager and Success Text Manager
Adds several new scripts to handle dialogue, and a scene with a Dialogue manager that handles both the Dialogue Manager and Success Text Manager scripts. The scene contains a button to test Dialogue Manager, which brings up a prototype dialogue box, as well as a button to test Success Text Manager, which plays an animation with a random phrase.